Move csvToGeoJson to CartoLayer
